Moved methods used by the public API classes from CacheSupport to CacheKeyGenerator; added test cases
