+++ httpclient5-cache/src/main/java/org/apache/hc/client5/http/impl/cache/BasicHttpCache.java
@@ -344,7 +344,7 @@ class BasicHttpCache implements HttpCache {
                 !Method.isSafe(request.getMethod())) {
             final String rootKey = cacheKeyGenerator.generateKey(host, request);
             evict(rootKey);
-            final URI requestUri = CacheSupport.normalize(CacheSupport.getRequestUri(request, host));
+            final URI requestUri = CacheKeyGenerator.normalize(CacheKeyGenerator.getRequestUri(host, request));
             if (requestUri != null) {
                 final URI contentLocation = CacheSupport.getLocationURI(requestUri, response, HttpHeaders.CONTENT_LOCATION);
                 if (contentLocation != null && CacheSupport.isSameOrigin(requestUri, contentLocation)) {
