+++ httpclient5-cache/src/main/java/org/apache/hc/client5/http/impl/cache/CacheKeyGenerator.java
@@ -38,6 +38,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.apache.hc.client5.http.cache.HttpCacheEntry;
 import org.apache.hc.core5.annotation.Contract;
+import org.apache.hc.core5.annotation.Internal;
 import org.apache.hc.core5.annotation.ThreadingBehavior;
 import org.apache.hc.core5.function.Resolver;
 import org.apache.hc.core5.http.Header;
@@ -45,7 +46,10 @@ import org.apache.hc.core5.http.HttpHeaders;
 import org.apache.hc.core5.http.HttpHost;
 import org.apache.hc.core5.http.HttpRequest;
 import org.apache.hc.core5.http.MessageHeaders;
+import org.apache.hc.core5.http.URIScheme;
 import org.apache.hc.core5.net.PercentCodec;
+import org.apache.hc.core5.net.URIAuthority;
+import org.apache.hc.core5.net.URIBuilder;
 import org.apache.hc.core5.util.Args;
 
 /**
@@ -62,6 +66,91 @@ public class CacheKeyGenerator implements Resolver<URI, String> {
     }
 
     /**
+     * Returns text representation of the request URI of the given {@link HttpRequest}.
+     * This method will use {@link HttpRequest#getPath()}, {@link HttpRequest#getScheme()} and
+     * {@link HttpRequest#getAuthority()} values when available or attributes of target
+     * {@link HttpHost } in order to construct an absolute URI.
+     * <p>
+     * This method will not attempt to ensure validity of the resultant text representation.
+     *
+     * @param target target host
+     * @param request the {@link HttpRequest}
+     *
+     * @return String the request URI
+     */
+    @Internal
+    public static String getRequestUri(final HttpHost target, final HttpRequest request) {
+        Args.notNull(target, "Target");
+        Args.notNull(request, "HTTP request");
+        final StringBuilder buf = new StringBuilder();
+        final URIAuthority authority = request.getAuthority();
+        if (authority != null) {
+            final String scheme = request.getScheme();
+            buf.append(scheme != null ? scheme : URIScheme.HTTP.id).append("://");
+            buf.append(authority.getHostName());
+            if (authority.getPort() >= 0) {
+                buf.append(":").append(authority.getPort());
+            }
+        } else {
+            buf.append(target.getSchemeName()).append("://");
+            buf.append(target.getHostName());
+            if (target.getPort() >= 0) {
+                buf.append(":").append(target.getPort());
+            }
+        }
+        final String path = request.getPath();
+        if (path == null) {
+            buf.append("/");
+        } else {
+            if (buf.length() > 0 && !path.startsWith("/")) {
+                buf.append("/");
+            }
+            buf.append(path);
+        }
+        return buf.toString();
+    }
+
+    /**
+     * Returns normalized representation of the request URI optimized for use as a cache key.
+     * This method ensures the resultant URI has an explicit port in the authority component,
+     * and explicit path component and no fragment.
+     */
+    @Internal
+    public static URI normalize(final URI requestUri) throws URISyntaxException {
+        Args.notNull(requestUri, "URI");
+        final URIBuilder builder = new URIBuilder(requestUri);
+        if (builder.getHost() != null) {
+            if (builder.getScheme() == null) {
+                builder.setScheme(URIScheme.HTTP.id);
+            }
+            if (builder.getPort() <= -1) {
+                if (URIScheme.HTTP.same(builder.getScheme())) {
+                    builder.setPort(80);
+                } else if (URIScheme.HTTPS.same(builder.getScheme())) {
+                    builder.setPort(443);
+                }
+            }
+        }
+        builder.setFragment(null);
+        return builder.normalizeSyntax().build();
+    }
+
+    /**
+     * Lenient URI parser that normalizes valid {@link URI}s and returns {@code null} for malformed URIs.
+     */
+    @Internal
+    public static URI normalize(final String requestUri) {
+        if (requestUri == null) {
+            return null;
+        }
+        try {
+            return CacheKeyGenerator.normalize(new URI(requestUri));
+        } catch (final URISyntaxException ex) {
+            return null;
+        }
+    }
+
+    /**
      * Computes a key for the given request {@link URI} that can be used as
      * a unique identifier for cached resources. The URI is expected to
      * in an absolute form.
@@ -71,7 +160,7 @@ public class CacheKeyGenerator implements Resolver<URI, String> {
      */
     public String generateKey(final URI requestUri) {
         try {
-            final URI normalizeRequestUri = CacheSupport.normalize(requestUri);
+            final URI normalizeRequestUri = normalize(requestUri);
             return normalizeRequestUri.toASCIIString();
         } catch (final URISyntaxException ex) {
             return requestUri.toASCIIString();
@@ -87,7 +176,7 @@ public class CacheKeyGenerator implements Resolver<URI, String> {
      * @return cache key
      */
     public String generateKey(final HttpHost host, final HttpRequest request) {
-        final String s = CacheSupport.getRequestUri(request, host);
+        final String s = getRequestUri(host, request);
         try {
             return generateKey(new URI(s));
         } catch (final URISyntaxException ex) {
