+++ httpclient5-cache/src/main/java/org/apache/hc/client5/http/impl/cache/CacheSupport.java
@@ -27,7 +27,6 @@
 package org.apache.hc.client5.http.impl.cache;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.BitSet;
 import java.util.Objects;
 import java.util.function.Consumer;
@@ -36,13 +35,8 @@ import org.apache.hc.client5.http.utils.URIUtils;
 import org.apache.hc.core5.annotation.Internal;
 import org.apache.hc.core5.http.FormattedHeader;
 import org.apache.hc.core5.http.Header;
-import org.apache.hc.core5.http.HttpHost;
-import org.apache.hc.core5.http.HttpRequest;
 import org.apache.hc.core5.http.MessageHeaders;
-import org.apache.hc.core5.http.URIScheme;
 import org.apache.hc.core5.http.message.ParserCursor;
-import org.apache.hc.core5.net.URIAuthority;
-import org.apache.hc.core5.net.URIBuilder;
 import org.apache.hc.core5.util.Args;
 import org.apache.hc.core5.util.CharArrayBuffer;
 import org.apache.hc.core5.util.TextUtils;
@@ -56,90 +50,6 @@ import org.apache.hc.core5.util.Tokenizer;
  */
 @Internal
 public final class CacheSupport {
-
-    private static final URI BASE_URI = URI.create("http://example.com/");
-
-    /**
-     * Returns text representation of the request URI of the given {@link HttpRequest}.
-     * This method will use {@link HttpRequest#getPath()}, {@link HttpRequest#getScheme()} and
-     * {@link HttpRequest#getAuthority()} values when available or attributes of target
-     * {@link HttpHost } in order to construct an absolute URI.
-     * <p>
-     * This method will not attempt to ensure validity of the resultant text representation.
-     *
-     * @param request the {@link HttpRequest}
-     * @param target target host
-     *
-     * @return String the request URI
-     */
-    public static String getRequestUri(final HttpRequest request, final HttpHost target) {
-        Args.notNull(request, "HTTP request");
-        Args.notNull(target, "Target");
-        final StringBuilder buf = new StringBuilder();
-        final URIAuthority authority = request.getAuthority();
-        if (authority != null) {
-            final String scheme = request.getScheme();
-            buf.append(scheme != null ? scheme : URIScheme.HTTP.id).append("://");
-            buf.append(authority.getHostName());
-            if (authority.getPort() >= 0) {
-                buf.append(":").append(authority.getPort());
-            }
-        } else {
-            buf.append(target.getSchemeName()).append("://");
-            buf.append(target.getHostName());
-            if (target.getPort() >= 0) {
-                buf.append(":").append(target.getPort());
-            }
-        }
-        final String path = request.getPath();
-        if (path == null) {
-            buf.append("/");
-        } else {
-            if (buf.length() > 0 && !path.startsWith("/")) {
-                buf.append("/");
-            }
-            buf.append(path);
-        }
-        return buf.toString();
-    }
-
-    /**
-     * Returns normalized representation of the request URI optimized for use as a cache key.
-     * This method ensures the resultant URI has an explicit port in the authority component,
-     * and explicit path component and no fragment.
-     */
-    public static URI normalize(final URI requestUri) throws URISyntaxException {
-        Args.notNull(requestUri, "URI");
-        final URIBuilder builder = new URIBuilder(requestUri.isAbsolute() ? URIUtils.resolve(BASE_URI, requestUri) : requestUri) ;
-        if (builder.getHost() != null) {
-            if (builder.getScheme() == null) {
-                builder.setScheme(URIScheme.HTTP.id);
-            }
-            if (builder.getPort() <= -1) {
-                if (URIScheme.HTTP.same(builder.getScheme())) {
-                    builder.setPort(80);
-                } else if (URIScheme.HTTPS.same(builder.getScheme())) {
-                    builder.setPort(443);
-                }
-            }
-        }
-        builder.setFragment(null);
-        return builder.normalizeSyntax().build();
-    }
-
-    /**
-     * Lenient URI parser that normalizes valid {@link URI}s and returns {@code null} for malformed URIs.
-     */
-    public static URI normalize(final String requestUri) {
-        if (requestUri == null) {
-            return null;
-        }
-        try {
-            return normalize(new URI(requestUri));
-        } catch (final URISyntaxException ex) {
-            return null;
-        }
-    }
 
     private static final BitSet COMMA = Tokenizer.INIT_BITSET(',');
 
@@ -179,7 +89,7 @@ public final class CacheSupport {
         if (h == null) {
             return null;
         }
-        final URI locationUri = CacheSupport.normalize(h.getValue());
+        final URI locationUri = CacheKeyGenerator.normalize(h.getValue());
         if (locationUri == null) {
             return requestUri;
         }
